Improved player movement, attacks, blocking. Added enemy script to skeleton.
The player movements and attacks are mapped to the keys and play animations when pressed. The skeleton is able to take damage from the player but will continue to keep playing the death animation when it dies. This bug needs to be fixed so we can use this script for other enemies.